Renamed file, moved import and variable declarations.
