+++ ts/Series/Bullet/BulletTargetOptions.d.ts
